LayoutTests/fast/frames/viewsource-linebreak.html

<html>
<head>
<script>
    if (window.testRunner) {
        testRunner.dumpAsText();
        testRunner.dumpChildFramesAsText();
    }
</script>
</head>
<body>
<p>You should see a frame in 'view source' mode below.</p>
<p>There should be no additional space at end of line. Number of line breaks should match source file.</p>
<hr>
<iframe viewsource name="sourceFrame" src="resources/viewsource-frame-5.html"></iframe>
</body>
</html>
