LayoutTests/fast/frames/viewsource-linebreak-expected.txt

You should see a frame in 'view source' mode below.

There should be no additional space at end of line. Number of line breaks should match source file.



--------
Frame: 'sourceFrame'
--------


<p>

<input


type="text">
</p>
